just stuff

+++ GamePrototype/Assets/Scripts/Coin.cs
@@ -7,7 +7,7 @@ public class Coin : Item {
     public override void OnPlayerPickup(Player p)
     {
         p.EnqueueFloatingText("+1 Gold", Color.yellow);
-        GameManager.S.AddGold(value);
+        GameManager.S.AddGold((int)value);
         Destroy(gameObject);
     }
 }
