+++ GamePrototype/Assets/Scripts/HealthPotion.cs
@@ -6,7 +6,7 @@ public class HealthPotion : Item {
 
     public override void OnPlayerPickup(Player p)
     {
-        p.currentHealth += value;
+        p.currentHealth += (int)value;
         p.UpdateHealthBar();
         p.EnqueueFloatingText(value + " Health restored!", Color.green);
         Destroy(gameObject);
